Update the example to use the new pop view controller when dragged right feature.
The example now uses the "pop view controller when dragged to the right" feature; it also handles the delegate methods to provide visual feedback to the user similar to the twitter app.

The example uses the SVPopOptionAllButFirst type (functions the same as twitter); this can be changed by changing the value of the popOffType property of the stack controller instance.